Registering two methods in NC. Documentation of neurite_util methods.

+++ test/test_neurite_proj.cpp
@@ -3983,7 +3983,7 @@ void create_spline_data_for_neurites
 		number erScaleFactor,
 		bool withER,
 		number anisotropy,
-		uint numRefs) {
+		size_t numRefs) {
 	{
 		// read in file to intermediate structure
 		std::vector<SWCPoint> vPoints;
